// frontend/src/app/models/certificado.model.ts
export class Certificado{
    id: string;
    curso: string;
    codigo: string;
    fecha_inicio: string;
    fecha_fin: string;
    horas: string;
    year: string;
    temario: string;
    carpeta: string;
    post_ruta: string;

    constructor(){
        this.id = "";
        this.curso = "";
        this.codigo = "";
        this.fecha_inicio = "";
        this.fecha_fin = "";
        this.horas = "";
        this.year = "";
        this.temario = "";
        this.carpeta = "";
        this.post_ruta = "";
    }
}